requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
openpyxl
pandas
python-dotenv
faiss-cpu
matplotlib
scikit-learn